+++ client/CHANGELOG.md
@@ -8,7 +8,7 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
-- Changed objects/works rights info ([#11](https://git.yale.edu/lux-its/lux-project/issues/11)).
+- Changed objects/works rights info ([#11](https://github.com/project-lux/lux-frontend/issues/11)).
 
 ### Removed